refactoring findrequest to cav_home and newrequest to pin_home.  adding message for no requests to list view
