+++ web-client/.gitignore
@@ -33,3 +33,6 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 package-lock.json
+
+yarn.lock
+.idea/
